kmd-weather-backend/requirements.txt
# Django and Core
Django==6.0
djangorestframework==3.16.1
django-environ==0.12.0
django-cors-headers==4.9.0
django-extensions==4.1
dj-database-url==3.1.0

# Database
psycopg2-binary==2.9.10

# Web Server
gunicorn==23.0.0
whitenoise==6.8.2

# Celery (Task Queue)
celery==5.6.0
redis==5.2.1

# SSH/SFTP
paramiko==4.0.0
scp==0.15.0
bcrypt==5.0.0
cryptography==46.0.3
PyNaCl==1.6.1

# GRIB Processing
pygrib==2.1.8
numpy==2.4.0
xarray==2025.12.0
pyproj==3.7.2

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzlocal==5.3.1